unsigned integers

+++ MapReduce/src/backend.cpp
@@ -72,7 +72,7 @@ void mapChunks(char* input, int length, std::unordered_map<std::string, int>* bu
 	#pragma omp parallel for
 	for (int i = 0; i < size; i++) {
 		for (int j = 0; j < MAX_THREADS; j++) {
-			for (int k = 0; k < localValues[i][j].size(); k++) {
+			for (long unsigned int k = 0; k < localValues[i][j].size(); k++) {
 				Pair tup = localValues[i][j].at(k);
 				if (buckets[i].find(tup.key) != buckets[i].end()) {
 					//key already exists. reduce locally
@@ -123,7 +123,7 @@ void mapReduce() {
 		uint64_t full_iterations = file_size / (size * chunk_size);
 		uint64_t leftover = file_size - full_iterations * (size * chunk_size);
 		
-		for(int i = 0; i < full_iterations; i++) {
+		for(uint64_t i = 0; i < full_iterations; i++) {
 			MPI_File_read_all(file, chunk, chunk_size, MPI_CHAR, MPI_STATUS_IGNORE);
 			
 			mapChunks(chunk, chunk_size, buckets);
@@ -307,7 +307,7 @@ void mapReduce() {
 	
 	std::unordered_map<std::string, int> map;
 	
-	for (int i = 0; i < received.size(); i++) {
+	for (long unsigned int i = 0; i < received.size(); i++) {
 		Pair tup = received[i];
 		if (map.find(tup.key) == map.end()) {
 			map.insert(make_pair(tup.key, tup.value));
